Increase JPEG size
Doubles the size of the requested JPEG and resizes modal window to 95% of the screen. There's not a 'good' way to predetermine a file size request from the client with the colorbox library, but I believe this is a decent compromise to produce a larger JPEG for users to preview layers.

fixes #22